test mixed

+++ src/fit-utils.jl
@@ -221,7 +221,7 @@ end
 function update_hist!(
     ::Type{L},
     hist::Vector{Matrix{Float64}},
-    ∇::Matrix{Float64},
+    ∇::Matrix{Float32},
     x_bin::Matrix,
     is::AbstractVector,
     js::AbstractVector,
@@ -244,7 +244,7 @@ end
 function update_hist!(
     ::Type{L},
     hist::Vector{Matrix{Float64}},
-    ∇::Matrix{Float64},
+    ∇::Matrix{Float32},
     x_bin::Matrix,
     is::AbstractVector,
     js::AbstractVector,
@@ -269,7 +269,7 @@ end
 function update_hist!(
     ::Type{L},
     hist::Vector{Matrix{Float64}},
-    ∇::Matrix{Float64},
+    ∇::Matrix{Float32},
     x_bin::Matrix,
     is::AbstractVector,
     js::AbstractVector,
